restructure db submodule
