+++ src/modules/osubot/db/index.js
@@ -1,0 +1,4 @@
+import userdb from './user'
+import statdb from './stat'
+import managedb from './manage'
+export { userdb, statdb, managedb }
