+++ src/modules/osubot/db/manage.js
@@ -1,0 +1,24 @@
+import fs from 'fs'
+import Monk from 'monk'
+import yaml from 'js-yaml'
+
+const config = yaml.safeLoad(fs.readFileSync('config.yml')).osubot
+const db = Monk(`localhost:${config.databasePort}/botdb`)
+const users = db.get('users')
+const usersBackup = db.get('backup')
+
+async function backup() {
+    const values = await users.find()
+    await usersBackup.remove({})
+    for (let user of values)
+        await usersBackup.insert(user)
+}
+
+async function recovery() {
+    const values = await usersBackup.find()
+    await users.remove({})
+    for (let user of values)
+        await users.insert(user)
+}
+
+export default { backup, recovery }
